Added NAN option to runs with negligible charge

+++ Tools/LibEnsemble/read_sim_output.py
@@ -37,33 +37,39 @@ def _beam_properties(filepath):
     ds = yt.load(filepath)
     ad = ds.all_data()
     w = ad['beam', 'particle_weight'].v
-    x = ad['beam', 'particle_position_x'].v
-    z = ad['beam', 'particle_position_y'].v
-    ux = ad['beam', 'particle_momentum_x'].v/scc.m_e/scc.c
-    uy = ad['beam', 'particle_momentum_y'].v/scc.m_e/scc.c
-    uz = ad['beam', 'particle_momentum_z'].v/scc.m_e/scc.c
+    if (w.shape[0] <= 200):
+        print('No particles in ',filepath)
+        return 0.0, 0.0, 0.0, 0.0
+    else:
+        x = ad['beam', 'particle_position_x'].v
+        z = ad['beam', 'particle_position_y'].v
+        ux = ad['beam', 'particle_momentum_x'].v/scc.m_e/scc.c
+        uy = ad['beam', 'particle_momentum_y'].v/scc.m_e/scc.c
+        uz = ad['beam', 'particle_momentum_z'].v/scc.m_e/scc.c
 
-    # Compute beam parameters
-    # Defined like that, the beam charge is > 0.
-    charge = np.sum(w) * scc.e
-    gamma = np.sqrt(1. + ux**2 + uy**2 + uz**2)
-    beta = np.sqrt(1.0 - 1.0 / gamma**2)
-    energy_MeV = scc.physical_constants["electron mass energy equivalent in MeV"][0] * (gamma - 1.)
-    energy_avg = np.average(energy_MeV, weights = w)
-    energy_std = np.average((energy_MeV - energy_avg)**2, weights = w) / energy_avg
-    nslices = 20
-    zslices = np.linspace(np.min(z), np.max(z), nslices+1)
-    exlist = np.zeros(nslices)
-    for slicei in range(nslices):
-        cond = [(z > zslices[slicei]) & (z < zslices[slicei+1])][0]
-        xslice = x[cond]
-        wslice = w[cond]
-        xpslice = ux[cond]/uz[cond]
-        gslice = gamma[cond]
-        bslice = beta[cond]
-        exlist[slicei] = slice_emittance(xslice, xpslice, gslice, bslice, wslice)
-    emittance = np.mean(exlist)
-    return charge, energy_avg, energy_std, emittance
+        # Compute beam parameters
+        # Defined like that, the beam charge is > 0.
+        charge = np.sum(w) * scc.e
+        gamma = np.sqrt(1. + ux**2 + uy**2 + uz**2)
+        beta = np.sqrt(1.0 - 1.0 / gamma**2)
+        energy_MeV = scc.physical_constants["electron mass energy equivalent in MeV"][0] * (gamma - 1.)
+        energy_avg = np.average(energy_MeV, weights = w)
+        energy_std = np.average((energy_MeV - energy_avg)**2, weights = w) / energy_avg
+        nslices = 20
+        zslices = np.linspace(np.min(z), np.max(z), nslices+1)
+        exlist = np.zeros(nslices)
+        for slicei in range(nslices):
+            cond = [(z > zslices[slicei]) & (z < zslices[slicei+1])][0]
+            wslice = w[cond]
+            if (wslice.shape[0] > 10):   
+                xslice = x[cond]
+                wslice = w[cond]
+                xpslice = ux[cond]/uz[cond]
+                gslice = gamma[cond]
+                bslice = beta[cond]
+                exlist[slicei] = slice_emittance(xslice, xpslice, gslice, bslice, wslice)
+        emittance = np.mean(exlist)
+        return charge, energy_avg, energy_std, emittance
 
 
 def read_sim_output(workdir):
@@ -86,14 +92,17 @@ def read_sim_output(workdir):
     filepath = os.path.join(workdir, datafile)
     charge_f, energy_avg, energy_std, emittance_f = _beam_properties(filepath)
 
-    # delete simulation results, just to have smaller data
-    shutil.rmtree('diags')
+    if (charge_f > 0.0):
+        # delete simulation results, just to have smaller data
+        shutil.rmtree('diags')
 
-    # Build a quantity to minimize (f) that encompasses
-    # emittance AND charge loss 1% charge loss has the
-    # same impact as doubling the initial emittance.
-    # we minimize f!
-    f = emittance_f + emittance_i*(1.-charge_f/charge_i)*100
-    warpx_out = np.array([f, energy_std, energy_avg, charge_f, emittance_f])
+        # Build a quantity to minimize (f) that encompasses
+        # emittance AND charge loss 1% charge loss has the
+        # same impact as doubling the initial emittance.
+        # we minimize f!
+        f = emittance_f + emittance_i*(1.-charge_f/charge_i)*100
+        warpx_out = np.array([f, energy_std, energy_avg, charge_f, emittance_f])
 
-    return warpx_out
+        return warpx_out
+    else:
+        return np.array([np.nan, np.nan, np.nan, np.nan, np.nan])
